requirements.txt
discord.py
APScheduler
numpy